+++ sofia_redux/toolkit/utilities/tests/test_func/test_faststack.py
@@ -9,6 +9,6 @@ def test_standard():
     y, x = np.mgrid[:2, :3]
     result = faststack(x, y)
     assert result.shape == (2, y.size)
-    assert np.issubdtype(np.int, result.dtype)
+    assert np.issubdtype(int, result.dtype)
     assert np.allclose(result[0], np.arange(6) % 3)
     assert np.allclose(result[1], np.arange(6) // 3)
